+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `prune-remote-emojis` command.
+- Prune remote emojis in background.
 
 ### Changed